refactor: split init setup steps
